MergeSort completed

+++ bubble_sort.h
@@ -3,7 +3,7 @@
 
 #include <iostream>
 
-void bubble_sort(int *array, size_t size) {
+void bubble_sort(int *array, int size) {
     bool swapped = true;
     while (swapped) {
         swapped = false;
